Special case handling of rspec-rails travis config
rspec-rails is ready to use the shared build config files. The only file
that needs to remain unique is the `.travis.yml` file. Special case the
handling for this file for the one repo.

Extract out logically grouped commands into descriptive method names to
aid in readability. Pass the pathnames all the way through the update
objects, so files are treated more like objects.