Better startup times through --dev
Opposed to what I said in rspec/rspec-core#2444 there is something
we can do, I just underestimated its effect. There is `--dev`
(see https://github.com/jruby/jruby/wiki/Improving-startup-time#use-the---dev-flag).

Locally for the rspec-core --bisect spec this brought the runtime
down from 33 to 36 seconds to 21 seconds!

It also encompasses disabling invoke dynamic already, so no need
for that. It is opposed to server mode that was used however,
server mode targets peak performance at the cost of bigger startup
and warmup times and is usually intended for running on the server/
benchmarks. It was introduced for no reason apparent to me in
bc0b97633ef00 and I think --dev fits the use case better and is
closer to what most peeps probably run either way.

Other changes I did:

* collapsed all the envs into one, it shouldn't hurt the other
  elements in the matrix, we lose the compat version 2.0 but
  that doesn't seem too important atm if it is please tell me
* updated the JRuby version to the most current one