Pin to minor versions
Having patch versions pinned comes with a burden of sending such pull
requests once in a while. We can depend on Travis to pick the latest
patch version as RVM updates.

On the other hand, problems with the patch versions would be discovered
later, during the next pull request.
This, however, is unlikely to happen too often.
Also, it should be possible to set up periodic builds for `master`.